+++ js/get-max-length-comment.js
@@ -1,0 +1,5 @@
+const COMMENT_LENGTH = 140;
+const getMaxLengthComment = function (comment) {
+  return (comment.length < COMMENT_LENGTH) ? true : false;
+};
+export {getMaxLengthComment};
